+++ README.md
@@ -4,27 +4,20 @@ Get GPT like chatGPT on your terminal
 
 ![Screenshot 2023-01-05 at 09 24 10](https://user-images.githubusercontent.com/11979969/210746185-69722c94-b073-4863-82bc-b662236c8305.png)
 
-
 # Get Started
 
 ```
-npm install
-```
-
-or
-
-```
-yarn
-```
-
-# Allow file
-
-```
-chmod +x ./bin/terminalGPT
+ npx terminalgpt
 ```
 
 # Run
 
 ```bash
-./bin/terminalGPT chat --engine "text-davinci-002" --temperature 0.7
+npx terminalgpt chat
+```
+
+# Changing engine and temperature
+
+```
+npx terminalgpt chat --engine "text-davinci-002" --temperature 0.7
 ```
