+++ package.json
@@ -1,7 +1,7 @@
 {
   "name": "terminalgpt",
-  "version": "1.0.0",
-  "main": "./bin/terminalGPT.js",
+  "version": "1.0.2",
+  "main": "index.js",
   "description": "Get GPT like chatGPT on your terminal",
   "scripts": {
     "test": "echo \"Error: no test specified\" && exit 1"
@@ -16,14 +16,19 @@
   "bugs": {
     "url": "https://github.com/jucasoliveira/terminalGPT/issues"
   },
-  "bin": "./bin/terminalGPT.js",
+  "bin": "bin/index.js",
   "homepage": "https://github.com/jucasoliveira/terminalGPT#readme",
   "dependencies": {
     "chalk": "^4.1.2",
     "commander": "^9.4.1",
+    "crypto": "^1.0.1",
     "gradient-string": "^2.0.2",
     "lowdb": "^5.0.5",
     "openai": "^3.1.0",
+    "ora": "^5.4.1",
     "prompts": "^2.4.2"
+  },
+  "devDependencies": {
+    "eslint": "^8.31.0"
   }
 }
